foundation for multiplayer logic
