+++ src/app/models/participant.spec.ts
@@ -1,0 +1,7 @@
+import { Participant } from './participant';
+
+describe('Participant', () => {
+  it('should create an instance', () => {
+    expect(new Participant()).toBeTruthy();
+  });
+});
