+++ src/app/models/score-board.spec.ts
@@ -1,0 +1,7 @@
+import { ScoreBoard } from './score-board';
+
+describe('ScoreBoard', () => {
+  it('should create an instance', () => {
+    expect(new ScoreBoard()).toBeTruthy();
+  });
+});
